ip_tracking/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-blacklist==0.7.0
django-ratelimit==4.1.0
sqlparse==0.5.3
